codelabs/trusted_space_codelab/src/requirements.txt
flask
google-cloud-storage
google-cloud-kms
torch
transformers
sentencepiece
cryptography